Split configuation classes file.
