+++ src/SFA.DAS.Commitments.Notification.WebJob/Configuration/NotificationsApiClientConfiguration.cs
@@ -1,0 +1,11 @@
+﻿using SFA.DAS.Notifications.Api.Client.Configuration;
+
+namespace SFA.DAS.Commitments.Notification.WebJob.Configuration
+{
+    public class NotificationsApiClientConfiguration : INotificationsApiClientConfiguration
+    {
+        public string BaseUrl { get; set; }
+
+        public string ClientToken { get; set; }
+    }
+}
