+++ src/SFA.DAS.Commitments.Notification.WebJob/SFA.DAS.Commitments.Notification.WebJob.csproj
@@ -150,7 +150,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DependencyResolution\CommitmentNotificationConfiguration.cs" />
+    <Compile Include="Configuration\AccountApiConfiguration.cs" />
+    <Compile Include="Configuration\CommitmentNotificationConfiguration.cs" />
+    <Compile Include="Configuration\NotificationsApiClientConfiguration.cs" />
     <Compile Include="DependencyResolution\DefaultRegistry.cs" />
     <Compile Include="DependencyResolution\FakeAccountApiClient.cs" />
     <Compile Include="DependencyResolution\IoC.cs" />
